работа над ошибками

+++ Sindie.ApiService/Sindie.ApiService.Core/Contracts/CreatureTemplateRequests/ChangeCreatureTemplate/ChangeCreatureTemplateRequestAppliedCondition.cs
@@ -20,6 +20,6 @@ namespace Sindie.ApiService.Core.Contracts.CreatureTemplateRequests.ChangeCreatu
 		/// <summary>
 		/// Шанс применения
 		/// </summary>
-		public double ApplyChance { get; set; }
+		public int ApplyChance { get; set; }
 	}
 }
